Sloppy Documentation. But Took 3 to 4 hours nevertheless Lost the point, of why i am documenting so much. Anyway when i look back, i m not going to be reading my notes. I will watch the course again or say.. google.
So, spending close to 4 hours on it, after every project seems pointless,
I can cover more course in that time.

1 point that i see, in trying to explain to someone, what i am doing is
trying to explain it to me. Otherwise, for namesakes i would have
completed a course, but i wud nothing about the beauty of it.

This is not a weak point, The same happened for me in DL.

I thought i knew everything.
Only when i started teaching, i knew i know nothing