CheckPoint
Same label names can appear in different functions of same file, or in different files.
We need to sum, filename + function_name to prefix such labels to avoid collision

Also, eq,gt,lt commands, create labels of their own. And since they are being created
in a automated way. They will carry the same names. So, u need the line_number.
But, then line-number is not unique between gts of 2 files.
So,we need to do Filename + Line_number

And this change need to work in both project 7 & 8