Project8 Completed. WEIRD.
Half the testcode doesn't work with Bootstrap code.
And Half the testcode works ONLY with Bootstrap code.

Bootstrap code, should ideally be setting SP to 256 & calling Sys.init.

But, here in the test files, SP is being set to 261 !!!

So, i need to check whether boostrap code shud be added or not.