Snake Complete
Detecting collision with the wall proved to be very difficult.
There was always some illegal index for screen error.

In the end, i made the snake go thru one edge and come thru the another.

BIGGEST Problem, that took me 2 whole days to figure out was...
Line 40 in SnakeCell.jack file
There is a boolean called if ((x = b.getX()) & (y = b.getY()))
Initially it had one set of brackets less .. like this
if (x = b.getX() & y = b.getY())

Since Jack doesnt use any BODMAS and priority of symbols during calculation..
I suffered a lot.

FINALLY DONE.

Only one bug.. since i give a sys.wait(200) so that the graphics is not too fast,
and the snake movement is not visible.. giving input to the system has become very difficult..

I have to hardpress for every single snake's movement..
This is because for some seconds, the system sleeps, during those seconds if i give input and leave key.. the system doesnt register it, and the key is never accepted..

I dont know whats the solution for that